requirements.txt
# VidChain v0.3.0 — Full dependency list
# Install GPU PyTorch separately:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Core Vision
opencv-python
ultralytics
torchvision
torchaudio
pillow<12.0

# Audio
openai-whisper
moviepy
imageio-ffmpeg
librosa
soundfile

# OCR
easyocr

# Emotion
deepface
tf-keras

# Scene Understanding (optional but recommended)
transformers

# RAG & Vector Store
sentence-transformers
chromadb
faiss-cpu

# LLM Routing
litellm
google-generativeai
python-dotenv

# Utilities
numpy
tqdm